svg/painting/parsing/stroke-miterlimit-computed.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: getComputedStyle().strokeMiterlimit</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#StrokeMiterlimitProperty" rel="help"/><h:meta content="stroke-miterlimit computed value is as specified." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/computed-testcommon.js"/><script><![CDATA[

test_computed_value("stroke-miterlimit", "0");
test_computed_value("stroke-miterlimit", "0.5");
test_computed_value("stroke-miterlimit", "1");
test_computed_value("stroke-miterlimit", "7.5");

  ]]></script></svg>
